Portfolio User işlemleri ViewModel validatorlar eklendi.

+++ App/App.AuthAPI/Controllers/UsersController.cs
@@ -121,7 +121,7 @@ public class UsersController : ControllerBase
         }
     }
 
-    [HttpPost("/edit-username")]
+    [HttpPut("/edit-username")]
     public async Task<IActionResult> EditUsernameAsync([FromBody] EditUsernameDto dto)
     {
         try
@@ -152,7 +152,7 @@ public class UsersController : ControllerBase
         }
     }
 
-    [HttpPost("/edit-user-image")]
+    [HttpPut("/change-user-image")]
     public async Task<IActionResult> EditUserImageAsync([FromBody] EditUserImageApiDto dto)
     {
         try
@@ -178,10 +178,14 @@ public class UsersController : ControllerBase
         }
     }
 
-
     [HttpDelete("/delete-user-img/{imgUrl}")]
     public async Task<IActionResult> DeleteUserImageAsync([FromRoute] string imgUrl)
     {
+        if (!string.IsNullOrEmpty(imgUrl))
+        {
+            return BadRequest(Result.Invalid());
+        }
+
         try
         {
             var result = await _userPortfolioService.DeleteUserImageAsync(imgUrl);
